revert: refactor of storageUsage to update the interface
This reverts commit 8ff98a407689da5036fe199e17fafb9a9d5131b8.
