+++ CHANGELOG.md
@@ -29,8 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- _(storageUsage)_ update the interface
-- _(fileTrasfer)_ add the file transfer interface
+- _(fileTransfer)_ add the file transfer interface
 
 ## [0.11.1] - 2026-05-27
