+++ Cargo.toml
@@ -207,7 +207,7 @@ systemd = { workspace = true, optional = true }
 udev = { workspace = true, optional = true }
 
 [target.'cfg(target_os = "windows")'.dependencies]
-windows-sys = { workspace = true, features = ["Win32_Storage_FileSystem"] }
+windows-sys = { workspace = true, features = ["Win32_Storage_FileSystem"], optional = true }
 
 [features]
 default = [
@@ -231,7 +231,7 @@ containers = [
 # installed
 cross = ["bindgen", "vendored"]
 # Enable the file transfer feature
-file-transfer = []
+file-transfer = ["windows-sys"]
 # Enable the forwarder service
 forwarder = ["dep:edgehog-forwarder"]
 # Enable connecting to astarte through the Message Hub
